refactor: clean uitl directory
폴더안으로 다 이동
