--- back/util/getScore.js
@@ -1,5 +1,0 @@
-function getScore(numerator, denominator, defaultScore) {
-  return Math.floor((numerator / denominator) * defaultScore);
-}
-
-module.exports = getScore;
